requirements.txt
requests>=2.24.0
python-dotenv>=0.13.0
scikit-learn>=0.23.1
iexfinance>=0.4.3
click>=7.1.2
